requirements.txt
fastapi
uvicorn
openai-agents
openai
python-dotenv
pydantic
